Improvements to buffer switcher with more config and custom fonts.
Buffer switcher improvements; support settings for short names, showing (or not) the number, and similar preferences to weechat (number character, max size for names and crop character), ability to specify a custom font.
Improved the preferences and added a font selection widget.
Added a seperate context menu for the toolbar to select icon label style.